Use puts instead of printf.
The classic "Hello World" example in C is actually often sort of wrong.
While printf(3) does the job, it is inefficient to use, since the
"format string" does not contain any format characters in this particular case.

puts(3) is better suited, since it does not try to interpret format characters,
and as an extra bonus, already appends the trailing newline for us.