Improve Whitespace
[The disassembly of the current program](https://gist.github.com/0x0dea/f68d381ed844f85fa2bd)
clearly indicates that it was generated by a machine, which is, I think,
not in the spirit of the project. This patch replaces the existing
program with a more idiomatic one which, in addition to being smaller
and faster, masquerades as a sequence of assembly instructions while
remaining an executable Whitespace program.