Fix c.c
Reverted back to the use of puts() since it is better for this case as
explained in commit 1b66868d5fd64bb6a677328764401fad96804ef3.

The reason given for the change from puts() to printf() in commit
52a8c332d87617f9e0ec59c6b81aad93afb1c220 was poor, and should not have
been accepted.

Functions that do not take a parameter in C must be declared with `void`
as its only parameter, otherwise you are describing a function with an
unspecified number of parameters.

Signed-off-by: Mahyar Koshkouei <mahyar.koshkouei@gmail.com>