Update and rename Vue.js to Vue.html
We include Vue.js from the CDN.
We create a <div> with an id of "app" where our Vue.js application will be mounted.
Inside the Vue instance, we define a data property called message and bind it to the {{ message }} expression in the template. The initial value of message is set to "Hello, World!".
When you open this HTML file in a web browser, it will display "Hello, World!" on the page. Vue.js takes care of rendering the message and updating it as needed.
This is a simple "Hello, World!" example using Vue.js to display a message on a webpage. You can later build more complex applications by adding components and handling user interactions.